Share common logic
